usersByID

+++ Goals/Goals/YNCListViewCell.h
@@ -14,6 +14,6 @@
 - (void)setTitle:(NSString *)title;
 - (void)setScore:(NSNumber *)score;
 - (void)setMembersWithImages:(NSArray *)images
-                       count:(int)count;
+                       count:(NSInteger)count;
 
 @end
